feat(ui): Update AddExpense component and modify related styles
The AddExpense component was heavily revised to improve the layout, user interactivity and information display. Notably, the ExpenseParties and CloseButton sub-components were introduced. Changes were made in the way friends and groups are rendered within the command list. This also necessitated adjustments in modal.css for consistency in style. Furthermore, unnecessary code in routeTree.gen.ts was removed.