Update codebase with path and schema tweaks
This commit includes changes in the tsconfig file by removing "@components/*" path. Modifications are also made in component import paths, replacing '@components/common' with '@/components/common'. In addition, there are revisions in the schemas.d.ts and routeTypes.d.ts files, such as the removal of unnecessary lines and changing double quotes to single quotes for consistency in code formatting.