feat (ui): Update UI components and enhance group activity tab
Improvement has been made to the user interface components, including the display of tabs and tabs' contents. The main focus is on the group activity tab, where most of the changes are applied to provide a more user-friendly interface. Additionally, the behavior of group expense view has been modified and grace handling of potentially null 'friend' object in GroupListItem has been added. Accordion animations are also introduced for a smoother user experience.