Refactor group membership syncing and creation flow
The code revision introduces changes concerning the organization and generation of group memberships. The Target interface was replaced with the more generic Object interface, adding better flexibility. The BulkCreateGroupMembership interface was removed and replaced with SyncGroupMembership, which syncs the group membership instead of creating it in bulk, reflecting an improved approach to group membership management. Furthermore, the user interface for creating groups was adjusted correspondingly.