feat (ui): Update expenses shares functionality in UI
This update refactors the ExpensesShares component and its associated functions in the UI. The changes also add a new optional property, 'onExpenseDetail' to handle expense details. Further improvements include the implementation of form validation and error handling as well as the addition of new class names for better UI responsiveness. This commit also makes changes to the order of the 'focusableInputElments' array and modifies the Currency component to accept an additional prop, 'noTabularNums'. The WatchState component now returns ReactNode instead of ReactElement.