Expense (#6)
* fix (ui): Add AddExpense modal and ParticipantsSelector UI component

The commit introduces a new 'AddExpense' modal and 'ParticipantsSelector' UI component. The 'AddExpense' modal is linked to the "Add Expense" button in the sidebar and bottom navigation. The 'ParticipantsSelector' component provides a functionality to search and select friends or groups for the expense. CSS styles for different elements in the listbox were also included.

* feat (ui): Add dynamic ExpenseItems component with FieldArray

Implemented a reusable FieldArray component, and integrated it into the ExpenseItems component. This helps support dynamic addition, removal, and rendering of expense items. Also, added the "just-throttle" package to manage the rate of operations and "@react-types/shared" for shared type definitions in the application.

* fix (ui): Remove unnecessary console.log from FieldArrayItems

Deleted a console.log statement from the onPress method inside FieldArrayItems.tsx. This change simplifies the onPress function to directly call the 'remove' function, making the code cleaner and easier to maintain.

* fix (ui): Update UI for dashboard groups and AddExpense modal

This commit makes several enhancements in the UI. It wraps the dashboard group link component in a div, adds an initial item count option to the FieldArray component and adds Back and Next buttons to the AddExpense modal for improved user navigation. Additionally, the form display style in the AddExpense modal is updated for improved layout.

* fix (ui): minor fixes

* feat (ui): Update ParticipantsSelector component

Refactored the ParticipantsSelector component with 'useParams' and 'useEffect' hooks for fetching both friend and group data. Also cleaned up the participant reducer logic by removing unnecessary code for item toggle. The user experience is also enhanced by early returns when group results or friends results do not exist yet.

* feat (ui): Add multi-step form and expense sharing feature

This update adds a multi-step form for the Add Expense modal. Field values and form context have been extended to scope field names, enabling better tracking of form progress and allowing sharing of expenses among users. Additionally, the Checkbox component has been updated to support different sizes and shapes and made more flexible for different functionalities with added attributes and styles.

* feat (ui): Implement drag-and-drop functionality for expense sharing

Added a new feature to dynamically allocate shares of an expense to different participants using a drag-and-drop interface. The implementation is based on a ListBox component and uses custom drag-and-drop Hooks. It simplifies the process of dividing an expense among various users within the Splinter app and improves user experience.

* feat (ui): add field transformers

* feat (ui): add field transformers

* feat (ui): updates form component

* feat (ui): initial add expense

* feat (ui): select all / none

* feat (ui): Update expenses shares functionality in UI

This update refactors the ExpensesShares component and its associated functions in the UI. The changes also add a new optional property, 'onExpenseDetail' to handle expense details. Further improvements include the implementation of form validation and error handling as well as the addition of new class names for better UI responsiveness. This commit also makes changes to the order of the 'focusableInputElments' array and modifies the Currency component to accept an additional prop, 'noTabularNums'. The WatchState component now returns ReactNode instead of ReactElement.