feat (ui): Add Alert component and animate form and field errors
Introduced a new Alert component into the common components library. This component will help in displaying various types of alerts across the app. Wrapper with animation is also added to error rendering in Form Fields and FormRootErrors for improved user experience. Color and style improvements are made in ButtonBase component.