fix: Adds support for typescript discriminated unions in PolymorphicSerializer OpenAPI Extension
A method, `build_typed_component`, was included in `PolymorphicSerializerExtension` to construct typed components based on conditions provided. It checks the Spectacular settings and based on that it defines the typed component name. Besides, it creates a schema and configures a new component if the sub-serializer has a non-empty scheme. The auto schema registry is updated and the method finally returns the created component.